+++ vec3_dot.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   vec3_dot.c                                         :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: jwalsh <marvin@42.fr>                      +#+  +:+       +#+        */
+/*   By: jwalsh <jwalsh@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/12/23 17:27:37 by jwalsh            #+#    #+#             */
-/*   Updated: 2017/02/06 13:35:45 by jwalsh           ###   ########.fr       */
+/*   Updated: 2017/02/12 15:17:44 by jwalsh           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -27,7 +27,7 @@
 ** gives you the angle a between the two vectors. (acos(x), x in radians)
 */
 
-float	vec3_dot(t_vec3 vec1, t_vec3 vec2)
+double	vec3_dot(t_vec3 vec1, t_vec3 vec2)
 {
 	return (vec1.x * vec2.x +
 			vec1.y * vec2.y +
